# Improve performance of `label2rgb`