--- a/benchmarks/skimage/cucim_color_bench.py
+++ b/benchmarks/skimage/cucim_color_bench.py
@@ -15,6 +15,7 @@ from _image_bench import ImageBench
 
 import cucim.skimage
 import cucim.skimage.color
+from cucim.skimage import data, measure
 
 func_name_choices = [
     "convert_colorspace",
@@ -88,7 +89,7 @@ class LabelBench(ImageBench):
         self,
         function_name,
         shape,
-        contiguous_labels=True,
+        image_none=False,
         dtypes=np.float32,
         fixed_kwargs={},
         var_kwargs={},
@@ -97,7 +98,7 @@ class LabelBench(ImageBench):
         module_gpu=cupyx.scipy.ndimage,
         run_cpu=True,
     ):
-        self.contiguous_labels = contiguous_labels
+        self.image_none = image_none
         super().__init__(
             function_name=function_name,
             shape=shape,
@@ -110,34 +111,40 @@ class LabelBench(ImageBench):
             run_cpu=run_cpu,
         )
 
+    def _generate_labels(self, dtype):
+        ndim = len(self.shape)
+        blobs_kwargs = dict(
+            blob_size_fraction=0.05, volume_fraction=0.35, rng=5
+        )
+        # binary blobs only creates square outputs
+        labels = measure.label(
+            data.binary_blobs(max(self.shape), n_dim=ndim, **blobs_kwargs)
+        )
+        print(f"# labels generated = {labels.max()}")
+
+        # crop to rectangular
+        labels = labels[tuple(slice(s) for s in self.shape)]
+        return labels.astype(dtype, copy=False)
+
     def set_args(self, dtype):
-        a = np.array(
-            [
-                [0, 0, 1, 1, 0, 0, 0, 0],
-                [0, 0, 0, 1, 0, 0, 4, 0],
-                [2, 2, 0, 0, 3, 0, 4, 4],
-                [0, 0, 0, 0, 0, 5, 0, 0],
-            ],
-            dtype=int,
-        )
-        tiling = tuple(s // a_s for s, a_s in zip(self.shape, a.shape))
-        if self.contiguous_labels:
-            label = np.kron(a, np.ones(tiling, dtype=a.dtype))
+        labels_d = self._generate_labels(dtype=np.int32)
+        labels = cp.asnumpy(labels_d)
+        if self.image_none:
+            self.args_cpu = (labels,)
+            self.args_gpu = (labels_d,)
         else:
-            label = np.tile(a, tiling)
-        labelled = cp.asarray(label)
-        imaged = cupy.testing.shaped_random(
-            labelled.shape, xp=cp, dtype=dtype, scale=1.0
-        )
-        image = cp.asnumpy(imaged)
-        self.args_cpu = (
-            label,
-            image,
-        )
-        self.args_gpu = (
-            labelled,
-            imaged,
-        )
+            imaged = cupy.testing.shaped_random(
+                labels.shape + (3,), xp=cp, dtype=dtype, scale=1.0
+            )
+            image = cp.asnumpy(imaged)
+            self.args_cpu = (
+                labels,
+                image,
+            )
+            self.args_gpu = (
+                labels_d,
+                imaged,
+            )
 
 
 def main(args):
@@ -225,19 +232,23 @@ def main(args):
             all_results = pd.concat([all_results, results["full"]])
 
         elif function_name == "label2rgb":
-            for contiguous_labels in [True, False]:
-                if contiguous_labels:
-                    index_str = "contiguous"
+            for image_none in [True, False]:
+                if image_none:
+                    # "avg" case cannot be used without an image
+                    var_kwargs = dict(kind=["overlay"])
+                    index_str = "without_image"
                 else:
-                    index_str = None
+                    var_kwargs = dict(kind=["avg", "overlay"])
+                    index_str = "with_image"
+
                 B = LabelBench(
                     function_name="label2rgb",
                     shape=shape,
                     dtypes=dtypes,
-                    contiguous_labels=contiguous_labels,
+                    image_none=image_none,
                     index_str=index_str,
                     fixed_kwargs=dict(bg_label=0),
-                    var_kwargs=dict(kind=["avg", "overlay"]),
+                    var_kwargs=var_kwargs,
                     module_cpu=skimage.color,
                     module_gpu=cucim.skimage.color,
                     run_cpu=run_cpu,