--- a/python/cucim/src/cucim/skimage/color/colorlabel.py
+++ b/python/cucim/src/cucim/skimage/color/colorlabel.py
@@ -1,4 +1,5 @@
 import itertools
+import math
 
 import cupy as cp
 import numpy as np
@@ -65,17 +66,17 @@ def _match_label_with_color(label, colors, bg_label, bg_color):
     # guarding against a potential corresponding future change for CuPy here
     mapped_labels = mapped_labels.reshape(-1)
 
-    # get rank of bg_label
-    bg_label_rank_list = mapped_labels[label.ravel() == bg_label]
-
     # The rank of each label is the index of the color it is matched to in
     # color cycle. bg_label should always be mapped to the first color, so
     # its rank must be 0. Other labels should be ranked from small to large
     # from 1.
-    if len(bg_label_rank_list) > 0:
-        bg_label_rank = bg_label_rank_list[0]
-        mapped_labels[mapped_labels < bg_label_rank] += 1
-        mapped_labels[label.ravel() == bg_label] = 0
+    bg_coords = (label.ravel() == bg_label).nonzero()
+    if bg_coords[0].size > 0:
+        # get rank of bg_label
+        bg_label_rank = mapped_labels[bg_coords[0][0]]
+        if bg_label_rank > 0:
+            mapped_labels[mapped_labels < bg_label_rank] += 1
+            mapped_labels[bg_coords] = 0
     else:
         mapped_labels += 1
 
@@ -144,7 +145,7 @@ def label2rgb(
         value in `label` with the image, at a certain alpha value.
     """
     if image is not None:
-        image = np.moveaxis(image, source=channel_axis, destination=-1)
+        image = cp.moveaxis(image, source=channel_axis, destination=-1)
     if kind == "overlay":
         rgb = _label2rgb_overlay(
             label,
@@ -161,6 +162,26 @@ def label2rgb(
     else:
         raise ValueError("`kind` must be either 'overlay' or 'avg'.")
     return np.moveaxis(rgb, source=-1, destination=channel_axis)
+
+
+alpha_scale_and_offset_ = cp.ElementwiseKernel(
+    "float64 alpha",
+    "X img1",
+    """
+    img1 = img1 * alpha + (1 - alpha);
+""",
+    name="alpha_scale_and_offset_",
+)
+
+
+alpha_blend_ = cp.ElementwiseKernel(
+    "X img1, Y img2, float64 alpha",
+    "Y result",
+    """
+    result = img1 * alpha + (1 - alpha) * img2;
+""",
+    name="alpha_scale_and_offset_",
+)
 
 
 def _label2rgb_overlay(
@@ -214,11 +235,7 @@ def _label2rgb_overlay(
         colors = DEFAULT_COLORS
     colors = [_rgb_vector(c) for c in colors]
 
-    if image is None:
-        image = cp.zeros(label.shape + (3,), dtype=np.float64)
-        # Opacity doesn't make sense if no image exists.
-        alpha = 1
-    else:
+    if image is not None:
         if (
             image.shape[: label.ndim] != label.shape
             or image.ndim > label.ndim + 1
@@ -233,13 +250,14 @@ def _label2rgb_overlay(
 
         float_dtype = _supported_float_type(image.dtype)
         image = img_as_float(image).astype(float_dtype, copy=False)
-        if image.ndim > label.ndim:
+        if image.ndim > label.ndim and saturation != 1.0:
             hsv = rgb2hsv(image)
             hsv[..., 1] *= saturation
             image = hsv2rgb(hsv)
         elif image.ndim == label.ndim:
             image = gray2rgb(image)
-        image = image * image_alpha + (1 - image_alpha)
+
+        alpha_scale_and_offset_(image_alpha, image)
 
     # Ensure that all labels are non-negative so we can index into
     # `label_to_color` correctly.
@@ -251,7 +269,7 @@ def _label2rgb_overlay(
     new_type = np.min_scalar_type(int(label.max()))
     if new_type == bool:
         new_type = np.uint8
-    label = label.astype(new_type)
+    label = label.astype(new_type, copy=False)
 
     mapped_labels_flat, color_cycle = _match_label_with_color(
         label, colors, bg_label, bg_color
@@ -269,14 +287,167 @@ def _label2rgb_overlay(
 
     mapped_labels = mapped_labels_flat.reshape(label.shape)
     label = mapped_labels
-    result = label_to_color[mapped_labels] * alpha + image * (1 - alpha)
 
-    # Remove background label if its color was not specified.
-    remove_background = 0 in mapped_labels_flat and bg_color is None
-    if remove_background:
-        result[label == bg_label] = image[label == bg_label]
+    if image is None:
+        result = label_to_color[mapped_labels]
+
+        # Remove background label if its color was not specified.
+        remove_background = 0 in mapped_labels_flat and bg_color is None
+        if remove_background:
+            result[label == bg_label] = 0
+    else:
+        result = alpha_blend_(label_to_color[mapped_labels], image, alpha)
+
+        # Remove background label if its color was not specified.
+        remove_background = 0 in mapped_labels_flat and bg_color is None
+        if remove_background:
+            result[label == bg_label] = image[label == bg_label]
 
     return result
+
+
+def _unravel_loop_index_declarations(var_name, ndim, uint_t="unsigned int"):
+    """Declare variables needed for unraveling the index to nd coordinates"""
+    if ndim == 1:
+        code = f"""
+        {uint_t} in_coord[1];"""
+        return code
+
+    code = f"""
+        // variables for unraveling a linear index to a coordinate array
+        {uint_t} in_coord[{ndim}];
+        {uint_t} temp_floor;"""
+    for d in range(ndim):
+        code += f"""
+        {uint_t} dim{d}_size = {var_name}.shape()[{d}];"""
+    return code
+
+
+def _unravel_loop_index(
+    var_name,
+    ndim,
+    uint_t="unsigned int",
+    raveled_index="i",
+    omit_declarations=False,
+):
+    """
+    declare a multi-index array in_coord and unravel the 1D index, i into it.
+    This code assumes that the array is a C-ordered array.
+    """
+    code = (
+        ""
+        if omit_declarations
+        else _unravel_loop_index_declarations(var_name, ndim, uint_t)
+    )
+    if ndim == 1:
+        code = f"""
+        in_coord[0] = {raveled_index};\n"""
+        return code
+
+    code += f"{uint_t} temp_idx = {raveled_index};"
+    for d in range(ndim - 1, 0, -1):
+        code += f"""
+        temp_floor = temp_idx / dim{d}_size;
+        in_coord[{d}] = temp_idx - temp_floor * dim{d}_size;
+        temp_idx = temp_floor;"""
+    code += """
+        in_coord[0] = temp_idx;"""
+    return code
+
+
+@cp.memoize(for_each_device=True)
+def get_roi_sums_and_counts_kernel(coord_dtype, pixels_per_thread=32):
+    coord_dtype = cp.dtype(coord_dtype)
+
+    # maximum number of unique labels this thread might encounter
+    # (could use a smaller value for most datasets, but for safety assume
+    # every pixel in the window could have a separate label.)
+    max_unique = pixels_per_thread
+
+    uint_t = (
+        "unsigned int" if coord_dtype.itemsize <= 4 else "unsigned long long"
+    )
+
+    # declare storage for local min/max of each label within the pixel window
+    source = f"""
+    {uint_t} start_index = {pixels_per_thread}*i;
+    float local_avg[{3 * pixels_per_thread}] = {{0}};\n
+    int local_count[{pixels_per_thread}] = {{0}};\n"""
+
+    # Find local average and count for each label among the next
+    # `pixels_per_thread` pixels
+    source += f"""
+      X encountered_labels[{max_unique}] = {{0}};
+      X current_label;
+      X prev_label = label[start_index];
+      int offset = 0;
+      encountered_labels[0] = prev_label;
+      {uint_t} ii_max = min(start_index + {pixels_per_thread}, label_size);
+      for ({uint_t} ii = start_index; ii < ii_max; ii++) {{
+        current_label = label[ii];
+        if (current_label == bg_label) {{ continue; }}
+        if (current_label != prev_label) {{
+            offset += 1;
+            prev_label = current_label;
+            encountered_labels[offset] = current_label;
+        }}
+        // inner loop operation for local_avg and offset
+        local_avg[3*offset] += static_cast<float>(img[3*ii]);
+        local_avg[3*offset + 1] += static_cast<float>(img[3*ii + 1]);
+        local_avg[3*offset + 2] += static_cast<float>(img[3*ii + 2]);
+        local_count[offset] += 1;
+      }}\n"""
+
+    # Update the global min/max values with the min/max from the pixel group
+    source += """
+      for (unsigned int ii = 0; ii <= offset; ii++) {
+        X lab = encountered_labels[ii];
+        if (lab != bg_label) {
+          atomicAdd(&avg[lab*3], local_avg[3*ii]);
+          atomicAdd(&avg[lab*3 + 1], local_avg[3*ii + 1]);
+          atomicAdd(&avg[lab*3 + 2], local_avg[3*ii + 2]);
+          atomicAdd(&count[lab], local_count[ii]);
+        }
+      }\n"""
+
+    inputs = "raw Y img, raw X label, "
+    inputs += f"raw {coord_dtype.name} label_size, raw int32 bg_label"
+    outputs = "raw float32 avg, raw int32 count"
+    name = f"cucim_roi_sums_and_counts__{coord_dtype.char}"
+    name += f"_batch{pixels_per_thread}"
+    return cp.ElementwiseKernel(inputs, outputs, source, name=name)
+
+
+# roi_sums_and_counts_ = cp.ElementwiseKernel(
+#     "raw Y img, X label, int32 bg_label",
+#     "raw float64 avg, raw int32 count",
+#     """
+#     if (label != bg_label) {
+#         atomicAdd(&avg[3*label], static_cast<double>(img[3*i]));
+#         atomicAdd(&avg[3*label + 1], static_cast<double>(img[3*i + 1]));
+#         atomicAdd(&avg[3*label + 2], static_cast<double>(img[3*i + 2]));
+#         atomicAdd(&count[label], 1);
+#     }
+# """,
+#     name="roi_sums_and_counts",
+# )
+
+roi_assign_averages_ = cp.ElementwiseKernel(
+    "raw float32 avg, Y label, int32 bg_label, raw X bg_color",
+    "raw X out",
+    """
+    if (label != bg_label) {
+        out[3*i] = avg[3*label];
+        out[3*i + 1] = avg[3*label + 1];
+        out[3*i + 2] = avg[3*label + 2];
+    } else {
+        out[3*i] = bg_color[0];
+        out[3*i + 1] = bg_color[1];
+        out[3*i + 2] = bg_color[2];
+    }
+""",
+    name="roi_assign_averages",
+)
 
 
 def _label2rgb_avg(label_field, image, bg_label=0, bg_color=(0, 0, 0)):
@@ -298,15 +469,61 @@ def _label2rgb_avg(label_field, image, bg_label=0, bg_color=(0, 0, 0)):
     out : ndarray, same shape and type as `image`
         The output visualization.
     """
-    out = cp.zeros(label_field.shape + (3,), dtype=image.dtype)
-    labels = cp.unique(label_field)
-    bg = labels == bg_label
-    if bg.any():
-        labels = labels[labels != bg_label]
-        mask = (label_field == bg_label).nonzero()
-        out[mask] = bg_color
-    for label in labels:
-        mask = (label_field == label).nonzero()
-        color = image[mask].mean(axis=0)
-        out[mask] = color
+    out = cp.empty(label_field.shape + (3,), dtype=image.dtype)
+    if image.ndim == label_field.ndim and image.shape == label_field.shape:
+        # convert to RGB as expected by roi_sums_and_counts_
+        float_dtype = _supported_float_type(image.dtype)
+        image = img_as_float(image).astype(float_dtype, copy=False)
+        image = gray2rgb(image)
+    if out.shape != image.shape:
+        raise ValueError(
+            f"{image.shape = }, but expected {out.shape} or {label_field.shape}"
+        )
+
+    # the kernels used assume C-order memory layout
+    image = cp.ascontiguousarray(image)
+    label_field = cp.ascontiguousarray(label_field)
+
+    # the following should be reasonable if the labels are sequential
+    # If the labels span a large range with many gaps, it may instead have been
+    # best to get a unique list of the labels first.
+    lmax = int(label_field.max())
+
+    # kernels assume label_field starts from zero so adjust it if necessary
+    lmin = int(label_field.min())
+    if lmin < 0:
+        label_field -= lmin
+        lmax -= lmin
+        bg_label -= lmin
+        lmin = 0
+
+    # choose 32 or 64-bit coordinate type for atomicMin and atomicMax
+    coord_dtype = cp.uint32 if image.size < 2**32 else cp.uint64
+
+    # Could potentially expose pixels per thread as a tuning parameter instead
+    # of using a fixed value here.
+    pixels_per_thread = 16
+    kernel = get_roi_sums_and_counts_kernel(
+        coord_dtype, pixels_per_thread=pixels_per_thread
+    )
+    avg = cp.zeros((lmax + 1, 3), dtype=cp.float32)
+    count = cp.zeros((lmax + 1, 1), dtype=cp.int32)
+    kernel(
+        image,
+        label_field,
+        label_field.size,
+        bg_label,
+        avg,
+        count,
+        size=math.ceil(label_field.size / pixels_per_thread),
+    )
+
+    # avoid divide by zero for background (average at bg_label not used)
+    if bg_label >= lmin and bg_label <= lmax:
+        count[bg_label] = 1
+    avg /= count
+
+    # assign the average or bg_color at each pixel in out
+    bg_color = cp.asarray(bg_color, dtype=out.dtype)
+    roi_assign_averages_(avg, label_field, bg_label, bg_color, out)
     return out